# JIT: contained memory safety analysis fixes and improvements