--- a/src/coreclr/jit/lower.h
+++ b/src/coreclr/jit/lower.h
@@ -567,7 +567,10 @@ public:
 
 #ifdef FEATURE_HW_INTRINSICS
     // Tries to get a containable node for a given HWIntrinsic
-    bool TryGetContainableHWIntrinsicOp(GenTreeHWIntrinsic* containingNode, GenTree** pNode, bool* supportsRegOptional);
+    bool TryGetContainableHWIntrinsicOp(GenTreeHWIntrinsic* containingNode,
+                                        GenTree**           pNode,
+                                        bool*               supportsRegOptional,
+                                        GenTreeHWIntrinsic* transparentParentNode = nullptr);
 #endif // FEATURE_HW_INTRINSICS
 
     static void TransformUnusedIndirection(GenTreeIndir* ind, Compiler* comp, BasicBlock* block);
@@ -586,6 +589,9 @@ private:
     // Checks for memory conflicts in the instructions between childNode and parentNode, and returns true if childNode
     // can be contained.
     bool IsSafeToContainMem(GenTree* parentNode, GenTree* childNode);
+
+    // Similar to above, but allows bypassing a "transparent" parent.
+    bool IsSafeToContainMem(GenTree* grandparentNode, GenTree* parentNode, GenTree* childNode);
 
     inline LIR::Range& BlockRange() const
     {