--- a/src/tests/JIT/opt/ForwardSub/lowerContainCheckCompare.csproj
+++ b/src/tests/JIT/opt/ForwardSub/lowerContainCheckCompare.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <DebugType />
+    <Optimize>True</Optimize>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildProjectName).cs" />
+  </ItemGroup>
+</Project>