… src/tests/JIT/opt/ForwardSub/andnotcontained.cs
… src/tests/JIT/opt/ForwardSub/lowerContainCheckCompare.cs