Fixes a couple of issues exposed by forward sub, where containment analysis
was allowing unsafe reordering of operands. More checks of this sort may be
needed.

See issues in #64828.

Generalize the safety check so that a store to a local not live into a handler
can be reordered with respect to node causing exceptions. Happily this leads
to almost uniformly better code despite the more stringent checking added above.

Add special handling for "transparent" nodes like CreateScalarUnsafe that allow
their children to be contained by their parent.

Add a workaround for the late callbacks into the containment checker made on
unlinked nodes. Assume these are always safe.