@AndyAyersMS 2022-02-05
Given that `Lowering::MakeSrcContained` takes the parent node as an arg it seems like we can add memory safety verification there and catch cases where the caller missed checks. 

Not sure if this will be super-costly for checked builds, but it seems worth trying.