@AndyAyersMS 2022-02-05
Small number of diffs from the extra checks. Here's one: we think we should be null checking `this` before bounds checking and this leads to some messy codegen.
```asm
;; BEFORE

; Assembly listing for method System.Runtime.Intrinsics.Vector128`1[Byte][System.Byte]:get_Item(int):ubyte:this
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No matching PGO data
; Final local variable assignments
;
;  V00 this         [V00,T01] (  3,  3   )   byref  ->  rcx         this single-def
;  V01 arg1         [V01,T00] (  4,  4   )     int  ->  rdx         single-def
;  V02 OutArgs      [V02    ] (  1,  1   )  lclBlk (32) [rsp+00H]   "OutgoingArgSpace"
;
; Lcl frame size = 40

G_M59860_IG01:        ; gcrefRegs=00000000 {}, byrefRegs=00000000 {}, byref, nogc <-- Prolog IG
       sub      rsp, 40
       vzeroupper 
						;; bbWeight=1    PerfScore 1.25
G_M59860_IG02:        ; gcrefRegs=00000000 {}, byrefRegs=00000002 {rcx}, byref, isz
       ; byrRegs +[rcx]
       cmp      edx, 16
       jae      SHORT G_M59860_IG04
       movzx    rax, byte  ptr [rcx+rdx]
						;; bbWeight=1    PerfScore 3.25
G_M59860_IG03:        ; , epilog, nogc, extend
       add      rsp, 40
       ret      
						;; bbWeight=1    PerfScore 1.25
G_M59860_IG04:        ; gcVars=0000000000000000 {}, gcrefRegs=00000000 {}, byrefRegs=00000000 {}, gcvars, byref
       ; byrRegs -[rcx]
       call     CORINFO_HELP_THROW_ARGUMENTOUTOFRANGEEXCEPTION
       ; gcr arg pop 0
       int3

;; AFTER

; Assembly listing for method System.Runtime.Intrinsics.Vector128`1[Byte][System.Byte]:get_Item(int):ubyte:this
; Emitting BLENDED_CODE for X64 CPU with AVX - Windows
; optimized code
; rsp based frame
; partially interruptible
; No matching PGO data
; Final local variable assignments
;
;  V00 this         [V00,T01] (  3,  3   )   byref  ->  rcx         this single-def
;  V01 arg1         [V01,T00] (  4,  4   )     int  ->  rdx         single-def
;  V02 OutArgs      [V02    ] (  1,  1   )  lclBlk (32) [rsp+00H]   "OutgoingArgSpace"
;  V03 rat0         [V03    ] (  1,  1   )  simd16  ->  [rsp+28H]   "SIMDInitTempVar"
;
; Lcl frame size = 56

G_M59860_IG01:        ; gcrefRegs=00000000 {}, byrefRegs=00000000 {}, byref, nogc <-- Prolog IG
       sub      rsp, 56
       vzeroupper 
						;; bbWeight=1    PerfScore 1.25
G_M59860_IG02:        ; gcrefRegs=00000000 {}, byrefRegs=00000002 {rcx}, byref, isz
       ; byrRegs +[rcx]
       vmovupd  xmm0, xmmword ptr [rcx]
       cmp      edx, 16
       jae      SHORT G_M59860_IG04
       vmovupd  xmmword ptr [rsp+28H], xmm0
       movzx    rax, byte  ptr [rsp+rdx+28H]
						;; bbWeight=1    PerfScore 8.25
G_M59860_IG03:        ; , epilog, nogc, extend
       add      rsp, 56
       ret      
						;; bbWeight=1    PerfScore 1.25
G_M59860_IG04:        ; gcVars=0000000000000000 {}, gcrefRegs=00000000 {}, byrefRegs=00000000 {}, gcvars, byref
       ; byrRegs -[rcx]
       call     CORINFO_HELP_THROW_ARGUMENTOUTOFRANGEEXCEPTION
       ; gcr arg pop 0
       int3  
```

Full set of regressions from the extra checks is

## aspnet.run.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 19733334 (overridden on cmd)
Total bytes of diff: 19733350 (overridden on cmd)
Total bytes of delta: 16 (0.00 % of base)
    diff is a regression.
    relative diff is a regression.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
           2 : 46659.dasm (1.49% of base)
           2 : 50687.dasm (0.32% of base)
           2 : 51092.dasm (1.49% of base)
           2 : 51969.dasm (0.35% of base)
           2 : 45076.dasm (1.49% of base)
           2 : 55877.dasm (1.49% of base)
           2 : 48677.dasm (0.37% of base)
           2 : 44556.dasm (0.32% of base)

8 total files with Code Size differences (0 improved, 8 regressed), 0 unchanged.

Top method regressions (bytes):
           2 ( 1.49% of base) : 46659.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 1.49% of base) : 51092.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 1.49% of base) : 45076.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 1.49% of base) : 55877.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 0.32% of base) : 50687.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool
           2 ( 0.35% of base) : 51969.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool
           2 ( 0.37% of base) : 48677.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool
           2 ( 0.32% of base) : 44556.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool

Top method regressions (percentages):
           2 ( 1.49% of base) : 46659.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 1.49% of base) : 51092.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 1.49% of base) : 45076.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 1.49% of base) : 55877.dasm - PathString:StartsWithSegments(PathString,int):bool:this
           2 ( 0.37% of base) : 48677.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool
           2 ( 0.35% of base) : 51969.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool
           2 ( 0.32% of base) : 50687.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool
           2 ( 0.32% of base) : 44556.dasm - RuntimeType:FilterApplyMethodBase(MethodBase,int,int,int,ref):bool

8 total methods with Code Size differences (0 improved, 8 regressed), 0 unchanged.

```

</details>

--------------------------------------------------------------------------------

## coreclr_tests.pmi.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 128321453 (overridden on cmd)
Total bytes of diff: 128321565 (overridden on cmd)
Total bytes of delta: 112 (0.00 % of base)
    diff is a regression.
    relative diff is a regression.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
          18 : 239319.dasm (3.01% of base)
          18 : 239317.dasm (2.95% of base)
          18 : 239318.dasm (2.85% of base)
          11 : 84869.dasm (42.31% of base)
          11 : 84868.dasm (39.29% of base)
          11 : 84867.dasm (40.74% of base)
          11 : 84896.dasm (40.74% of base)
          11 : 84870.dasm (39.29% of base)
           2 : 239321.dasm (0.32% of base)
           2 : 89419.dasm (0.09% of base)

Top file improvements (bytes):
          -1 : 239320.dasm (-0.16% of base)

11 total files with Code Size differences (1 improved, 10 regressed), 0 unchanged.

Top method regressions (bytes):
          18 ( 2.95% of base) : 239317.dasm - VectorArrayTest`1[Byte][System.Byte]:VectorArray(ubyte):int
          18 ( 2.85% of base) : 239318.dasm - VectorArrayTest`1[Int16][System.Int16]:VectorArray(short):int
          18 ( 3.01% of base) : 239319.dasm - VectorArrayTest`1[Int32][System.Int32]:VectorArray(int):int
          11 (40.74% of base) : 84867.dasm - System.Runtime.Intrinsics.Vector128`1[Byte][System.Byte]:get_Item(int):ubyte:this
          11 (39.29% of base) : 84870.dasm - System.Runtime.Intrinsics.Vector128`1[Double][System.Double]:get_Item(int):double:this
          11 (39.29% of base) : 84868.dasm - System.Runtime.Intrinsics.Vector128`1[Int16][System.Int16]:get_Item(int):short:this
          11 (42.31% of base) : 84869.dasm - System.Runtime.Intrinsics.Vector128`1[Int32][System.Int32]:get_Item(int):int:this
          11 (40.74% of base) : 84896.dasm - System.Runtime.Intrinsics.Vector128`1[Int64][System.Int64]:get_Item(int):long:this
           2 ( 0.09% of base) : 89419.dasm - ILGEN_0x537f7b0:Method_0x323f83b5(double,byte,int,short,ubyte,double,short,ushort,double,long,int,long,ushort,long,int):float
           2 ( 0.32% of base) : 239321.dasm - VectorArrayTest`1[Int64][System.Int64]:VectorArray(long):int

Top method improvements (bytes):
          -1 (-0.16% of base) : 239320.dasm - VectorArrayTest`1[Double][System.Double]:VectorArray(double):int

Top method regressions (percentages):
          11 (42.31% of base) : 84869.dasm - System.Runtime.Intrinsics.Vector128`1[Int32][System.Int32]:get_Item(int):int:this
          11 (40.74% of base) : 84867.dasm - System.Runtime.Intrinsics.Vector128`1[Byte][System.Byte]:get_Item(int):ubyte:this
          11 (40.74% of base) : 84896.dasm - System.Runtime.Intrinsics.Vector128`1[Int64][System.Int64]:get_Item(int):long:this
          11 (39.29% of base) : 84870.dasm - System.Runtime.Intrinsics.Vector128`1[Double][System.Double]:get_Item(int):double:this
          11 (39.29% of base) : 84868.dasm - System.Runtime.Intrinsics.Vector128`1[Int16][System.Int16]:get_Item(int):short:this
          18 ( 3.01% of base) : 239319.dasm - VectorArrayTest`1[Int32][System.Int32]:VectorArray(int):int
          18 ( 2.95% of base) : 239317.dasm - VectorArrayTest`1[Byte][System.Byte]:VectorArray(ubyte):int
          18 ( 2.85% of base) : 239318.dasm - VectorArrayTest`1[Int16][System.Int16]:VectorArray(short):int
           2 ( 0.32% of base) : 239321.dasm - VectorArrayTest`1[Int64][System.Int64]:VectorArray(long):int
           2 ( 0.09% of base) : 89419.dasm - ILGEN_0x537f7b0:Method_0x323f83b5(double,byte,int,short,ubyte,double,short,ushort,double,long,int,long,ushort,long,int):float

Top method improvements (percentages):
          -1 (-0.16% of base) : 239320.dasm - VectorArrayTest`1[Double][System.Double]:VectorArray(double):int

11 total methods with Code Size differences (1 improved, 10 regressed), 0 unchanged.

```

</details>

--------------------------------------------------------------------------------

## libraries.pmi.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 46583443 (overridden on cmd)
Total bytes of diff: 46583502 (overridden on cmd)
Total bytes of delta: 59 (0.00 % of base)
    diff is a regression.
    relative diff is a regression.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
          11 : 20673.dasm (40.74% of base)
          11 : 20675.dasm (42.31% of base)
          11 : 20674.dasm (39.29% of base)
          11 : 20676.dasm (39.29% of base)
          11 : 20701.dasm (40.74% of base)
           2 : 58129.dasm (0.41% of base)
           2 : 58130.dasm (0.71% of base)

7 total files with Code Size differences (0 improved, 7 regressed), 0 unchanged.

Top method regressions (bytes):
          11 (40.74% of base) : 20673.dasm - System.Runtime.Intrinsics.Vector128`1[Byte][System.Byte]:get_Item(int):ubyte:this
          11 (39.29% of base) : 20676.dasm - System.Runtime.Intrinsics.Vector128`1[Double][System.Double]:get_Item(int):double:this
          11 (39.29% of base) : 20674.dasm - System.Runtime.Intrinsics.Vector128`1[Int16][System.Int16]:get_Item(int):short:this
          11 (42.31% of base) : 20675.dasm - System.Runtime.Intrinsics.Vector128`1[Int32][System.Int32]:get_Item(int):int:this
          11 (40.74% of base) : 20701.dasm - System.Runtime.Intrinsics.Vector128`1[Int64][System.Int64]:get_Item(int):long:this
           2 ( 0.71% of base) : 58130.dasm - Microsoft.CodeAnalysis.VisualBasic.LocalRewriter:AppendToBlock(Microsoft.CodeAnalysis.VisualBasic.BoundBlock,Microsoft.CodeAnalysis.VisualBasic.BoundStatement):Microsoft.CodeAnalysis.VisualBasic.BoundBlock:this
           2 ( 0.41% of base) : 58129.dasm - Microsoft.CodeAnalysis.VisualBasic.LocalRewriter:Concat(Microsoft.CodeAnalysis.VisualBasic.BoundStatement,Microsoft.CodeAnalysis.VisualBasic.BoundStatement):Microsoft.CodeAnalysis.VisualBasic.BoundStatement:this

Top method regressions (percentages):
          11 (42.31% of base) : 20675.dasm - System.Runtime.Intrinsics.Vector128`1[Int32][System.Int32]:get_Item(int):int:this
          11 (40.74% of base) : 20673.dasm - System.Runtime.Intrinsics.Vector128`1[Byte][System.Byte]:get_Item(int):ubyte:this
          11 (40.74% of base) : 20701.dasm - System.Runtime.Intrinsics.Vector128`1[Int64][System.Int64]:get_Item(int):long:this
          11 (39.29% of base) : 20676.dasm - System.Runtime.Intrinsics.Vector128`1[Double][System.Double]:get_Item(int):double:this
          11 (39.29% of base) : 20674.dasm - System.Runtime.Intrinsics.Vector128`1[Int16][System.Int16]:get_Item(int):short:this
           2 ( 0.71% of base) : 58130.dasm - Microsoft.CodeAnalysis.VisualBasic.LocalRewriter:AppendToBlock(Microsoft.CodeAnalysis.VisualBasic.BoundBlock,Microsoft.CodeAnalysis.VisualBasic.BoundStatement):Microsoft.CodeAnalysis.VisualBasic.BoundBlock:this
           2 ( 0.41% of base) : 58129.dasm - Microsoft.CodeAnalysis.VisualBasic.LocalRewriter:Concat(Microsoft.CodeAnalysis.VisualBasic.BoundStatement,Microsoft.CodeAnalysis.VisualBasic.BoundStatement):Microsoft.CodeAnalysis.VisualBasic.BoundStatement:this

7 total methods with Code Size differences (0 improved, 7 regressed), 0 unchanged.

```

</details>

--------------------------------------------------------------------------------

## libraries_tests.pmi.windows.x64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 119636782 (overridden on cmd)
Total bytes of diff: 119637034 (overridden on cmd)
Total bytes of delta: 252 (0.00 % of base)
    diff is a regression.
    relative diff is a regression.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
          33 : 305492.dasm (6.53% of base)
          27 : 305489.dasm (5.76% of base)
          27 : 305491.dasm (5.81% of base)
          27 : 305494.dasm (5.72% of base)
          27 : 305504.dasm (5.57% of base)
          27 : 305490.dasm (5.73% of base)
          21 : 305503.dasm (4.67% of base)
          21 : 305502.dasm (4.61% of base)
          21 : 305501.dasm (4.64% of base)
          21 : 305506.dasm (4.59% of base)

10 total files with Code Size differences (0 improved, 10 regressed), 0 unchanged.

Top method regressions (bytes):
          33 ( 6.53% of base) : 305492.dasm - <>c__DisplayClass309_0`1[Double][System.Double]:<TestAdditionOverflow>b__0(int,double):this
          27 ( 5.76% of base) : 305489.dasm - <>c__DisplayClass309_0`1[Byte][System.Byte]:<TestAdditionOverflow>b__0(int,ubyte):this
          27 ( 5.73% of base) : 305490.dasm - <>c__DisplayClass309_0`1[Int16][System.Int16]:<TestAdditionOverflow>b__0(int,short):this
          27 ( 5.81% of base) : 305491.dasm - <>c__DisplayClass309_0`1[Int32][System.Int32]:<TestAdditionOverflow>b__0(int,int):this
          27 ( 5.72% of base) : 305494.dasm - <>c__DisplayClass309_0`1[Int64][System.Int64]:<TestAdditionOverflow>b__0(int,long):this
          27 ( 5.57% of base) : 305504.dasm - <>c__DisplayClass329_0`1[Double][System.Double]:<TestSubtractionOverflow>b__0(int,double):this
          21 ( 4.64% of base) : 305501.dasm - <>c__DisplayClass329_0`1[Byte][System.Byte]:<TestSubtractionOverflow>b__0(int,ubyte):this
          21 ( 4.61% of base) : 305502.dasm - <>c__DisplayClass329_0`1[Int16][System.Int16]:<TestSubtractionOverflow>b__0(int,short):this
          21 ( 4.67% of base) : 305503.dasm - <>c__DisplayClass329_0`1[Int32][System.Int32]:<TestSubtractionOverflow>b__0(int,int):this
          21 ( 4.59% of base) : 305506.dasm - <>c__DisplayClass329_0`1[Int64][System.Int64]:<TestSubtractionOverflow>b__0(int,long):this

Top method regressions (percentages):
          33 ( 6.53% of base) : 305492.dasm - <>c__DisplayClass309_0`1[Double][System.Double]:<TestAdditionOverflow>b__0(int,double):this
          27 ( 5.81% of base) : 305491.dasm - <>c__DisplayClass309_0`1[Int32][System.Int32]:<TestAdditionOverflow>b__0(int,int):this
          27 ( 5.76% of base) : 305489.dasm - <>c__DisplayClass309_0`1[Byte][System.Byte]:<TestAdditionOverflow>b__0(int,ubyte):this
          27 ( 5.73% of base) : 305490.dasm - <>c__DisplayClass309_0`1[Int16][System.Int16]:<TestAdditionOverflow>b__0(int,short):this
          27 ( 5.72% of base) : 305494.dasm - <>c__DisplayClass309_0`1[Int64][System.Int64]:<TestAdditionOverflow>b__0(int,long):this
          27 ( 5.57% of base) : 305504.dasm - <>c__DisplayClass329_0`1[Double][System.Double]:<TestSubtractionOverflow>b__0(int,double):this
          21 ( 4.67% of base) : 305503.dasm - <>c__DisplayClass329_0`1[Int32][System.Int32]:<TestSubtractionOverflow>b__0(int,int):this
          21 ( 4.64% of base) : 305501.dasm - <>c__DisplayClass329_0`1[Byte][System.Byte]:<TestSubtractionOverflow>b__0(int,ubyte):this
          21 ( 4.61% of base) : 305502.dasm - <>c__DisplayClass329_0`1[Int16][System.Int16]:<TestSubtractionOverflow>b__0(int,short):this
          21 ( 4.59% of base) : 305506.dasm - <>c__DisplayClass329_0`1[Int64][System.Int64]:<TestSubtractionOverflow>b__0(int,long):this

10 total methods with Code Size differences (0 improved, 10 regressed), 0 unchanged.

```

</details>

--------------------------------------------------------------------------------

Going to push this next batch of changes and will come back to addressing regressions when I have time.